[MODIFY] 구조 수정
