+++ src/main/java/com/project/bug_report/controller/ReviewController.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ReviewController {
+}
