--- src/main/java/com/project/bug_report/domain/company/CompanyRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.company;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-}
