--- src/main/java/com/project/bug_report/domain/company_reply/CompanyReplyRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.company_reply;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CompanyReplyRepository extends JpaRepository<CompanyReply, Long> {
-}
