--- src/main/java/com/project/bug_report/domain/nested_comment/NestedCommentRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.nested_comment;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NestedCommentRepository extends JpaRepository<NestedComment, Long> {
-}
