--- src/main/java/com/project/bug_report/domain/posts/PostsRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.posts;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostsRepository extends JpaRepository<Posts, Long> {
-}
