--- src/main/java/com/project/bug_report/domain/review/ReviewRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.review;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-}
