--- src/main/java/com/project/bug_report/domain/review_detail/ReviewDetailRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.review_detail;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReviewDetailRepository extends JpaRepository<ReviewDetail, Long> {
-}
