--- src/main/java/com/project/bug_report/domain/service/ServiceRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.service;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ServiceRepository extends JpaRepository<Service, Long> {
-}
