--- src/main/java/com/project/bug_report/domain/service_category/ServiceCategory.java
@@ -1,25 +1,0 @@
-package com.project.bug_report.domain.service_category;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@NoArgsConstructor
-@Entity
-public class ServiceCategory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long categoryId;
-
-    @Column
-    private String categoryName;
-
-    @Builder
-    public ServiceCategory(String categoryName){
-        this.categoryName = categoryName;
-    }
-}
