--- src/main/java/com/project/bug_report/domain/service_category/ServiceCategoryRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.service_category;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ServiceCategoryRepository extends JpaRepository<ServiceCategory, Long> {
-}
