--- src/main/java/com/project/bug_report/domain/user/UserRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
