--- src/main/java/com/project/bug_report/domain/user_comment/UserCommentRepository.java
@@ -1,6 +1,0 @@
-package com.project.bug_report.domain.user_comment;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserCommentRepository extends JpaRepository<UserComment, Long> {
-}
