+++ src/main/java/com/project/bug_report/model/product/ProductCategory.java
@@ -1,0 +1,26 @@
+package com.project.bug_report.model.product;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.io.Serializable;
+
+@Getter
+@NoArgsConstructor
+@Entity
+public class ProductCategory implements Serializable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long categoryId;
+
+    @Column
+    private String categoryName;
+
+    @Builder
+    public ProductCategory(String categoryName){
+        this.categoryName = categoryName;
+    }
+}
