+++ src/main/java/com/project/bug_report/repository/product/CompanyReplyRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.product;
+
+import com.project.bug_report.model.product.CompanyReply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CompanyReplyRepository extends JpaRepository<CompanyReply, Long> {
+}
