+++ src/main/java/com/project/bug_report/repository/product/ProductCategoryRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.product;
+
+import com.project.bug_report.model.product.ProductCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
+}
