+++ src/main/java/com/project/bug_report/repository/product/ProductRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.product;
+
+import com.project.bug_report.model.product.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
