+++ src/main/java/com/project/bug_report/repository/review/NestedCommentRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.review;
+
+import com.project.bug_report.model.review.NestedComment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NestedCommentRepository extends JpaRepository<NestedComment, Long> {
+}
