+++ src/main/java/com/project/bug_report/repository/review/ReviewDetailRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.review;
+
+import com.project.bug_report.model.review.ReviewDetail;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReviewDetailRepository extends JpaRepository<ReviewDetail, Long> {
+}
