+++ src/main/java/com/project/bug_report/repository/review/ReviewRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.review;
+
+import com.project.bug_report.model.review.Review;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+}
