+++ src/main/java/com/project/bug_report/repository/review/UserCommentRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.review;
+
+import com.project.bug_report.model.review.UserComment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserCommentRepository extends JpaRepository<UserComment, Long> {
+}
