+++ src/main/java/com/project/bug_report/repository/user/UserRepository.java
@@ -1,0 +1,7 @@
+package com.project.bug_report.repository.user;
+
+import com.project.bug_report.model.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
