+++ src/main/java/com/project/bug_report/wait4rm/BaseController.java
@@ -1,0 +1,13 @@
+//package com.project.bug_report.controller;
+//
+//import org.springframework.web.bind.annotation.GetMapping;
+//import org.springframework.web.bind.annotation.RestController;
+//
+//@RestController
+//public class BaseController {
+//
+//    @GetMapping("/")
+//    public String home(){
+//        return "Bug report service";
+//    }
+//}
