+++ src/main/java/com/project/bug_report/wait4rm/Posts.java
@@ -1,0 +1,33 @@
+//package com.project.bug_report.wait4rm;
+//
+//import lombok.Builder;
+//import lombok.Getter;
+//import lombok.NoArgsConstructor;
+//
+//import javax.persistence.*;
+//
+//@Getter
+//@NoArgsConstructor
+//@Entity
+//public class Posts {
+//
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private Long id;
+//
+//    @Column(nullable = false)
+//    private String title;
+//
+//    @Column(nullable = false)
+//    private String content;
+//
+//    @Column(nullable = false)
+//    private String author;
+//
+//    @Builder
+//    public Posts(String title, String content, String author){
+//        this.title = title;
+//        this.content = content;
+//        this.author = author;
+//    }
+//}
