+++ src/main/java/com/project/bug_report/wait4rm/PostsRepository.java
@@ -1,0 +1,6 @@
+//package com.project.bug_report.wait4rm;
+//
+//import org.springframework.data.jpa.repository.JpaRepository;
+//
+//public interface PostsRepository extends JpaRepository<Posts, Long> {
+//}
