--- src/main/java/com/project/bug_report/web/IndexController.java
@@ -1,13 +1,0 @@
-package com.project.bug_report.web;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class IndexController {
-
-    @GetMapping("/")
-    public String home(){
-        return "This is bug report application";
-    }
-}
